Allow arbitrary GTF/FASTA URLs
