+++ README.md
@@ -17,6 +17,9 @@ gene_names = data.gene_names_at_locus(contig=6, position=29945884)
 # get all exons associated with HLA-A
 exon_ids  = data.exon_ids_of_gene_name('HLA-A')
 ```
+# Arbitrary Genomic Databases
+
+PyEnsembl also allows arbitrary genomic databases as follows:
 
 # Installation
