+++ pyensembl/sequence_data.py
@@ -21,32 +21,36 @@ from Bio import SeqIO
 import datacache
 
 from .common import CACHE_SUBDIR, require_ensembl_id
-from .release_info import check_release_number
 
 
 class SequenceData(object):
-    """Container for reference nucleotide and amino acid sequenes.
+    """
+    Container for reference nucleotide and amino acid sequenes.
 
     Downloads and caches FASTA file, creates database mapping
     identifiers to sequences.
     """
     def __init__(
             self,
-            ensembl_release,
-            fasta_url,
+            genome_source,
+            fasta_type,
+            local_filename_func=None,
+            require_ensembl_ids=True,
             auto_download=False):
 
         # download cache for fetching reference FASTA files
         self.cache = datacache.Cache(CACHE_SUBDIR)
         self.auto_download = auto_download
-        self.release = check_release_number(ensembl_release)
         # expect the remote FASTA file to be gzipped and the local needs
         # to be a decompressed text file
         self.fasta_decompress = True
-        self.url = fasta_url
+        self.url = genome_source.fasta_url(fasta_type)
         self.remote_filename = split(self.url)[1]
 
-        self.local_filename = self._create_local_filename(self.remote_filename)
+        if local_filename_func:
+            self.local_filename = local_filename_func(self.remote_filename)
+        else:
+            self.local_filename = self.remote_filename
         self._init_lazy_fields()
 
     def _init_lazy_fields(self):
@@ -61,29 +65,8 @@ class SequenceData(object):
         # key set for fasta dictionary, for faster membership tests,
         self._fasta_keys = None
 
-    def _create_local_filename(self, remote_filename):
-        """
-        We sometimes need to add the release number to a local FASTAfilename
-        since some Ensembl releases only have the genome name in the FASTA
-        filename but still differ subtly between releases.
-        For example, a transcript ID may be missing in Ensembl 75 but present
-        in 76, though both have the same FASTA filename
-        """
-        if ".%d." % self.release in remote_filename:
-            return remote_filename
-
-        filename_parts = self.remote_filename.split(".fa.")
-        assert len(filename_parts) == 2, \
-            "Expected remote filename %s to contain '.fa.gz'" % (
-                self.remote_filename,)
-        return "".join([
-            filename_parts[0],
-            ".%d.fa." % self.release,
-            filename_parts[1]])
-
     def __str__(self):
-        return "SequenceData(ensembl_release=%d, url=%s, local_path=%s)" % (
-            self.release,
+        return "SequenceData(url=%s, local_path=%s)" % (
             self.url,
             self.local_fasta_path)
 
@@ -99,7 +82,7 @@ class SequenceData(object):
         return (
             isinstance(other, SequenceData) and
             self.url == other.url and
-            self.release == other.release)
+            self.version == other.version)
 
     @property
     def local_fasta_path(self):
@@ -115,11 +98,12 @@ class SequenceData(object):
             # Does a download if the cache is empty, otherwise just returns
             # the local path
             return self._fetch(force=False)
-        raise ValueError("Ensembl sequence data %s is not currently "
-                         "installed for release %s. Run "
-                         "\"pyensembl install --relase %s\" or call "
-                         "EnsemblRelease(%s).install()" %
-                         ((self.local_filename,) + (self.release,) * 3))
+        raise ValueError("Genome %s sequence data %s is not currently "
+                         "installed for this genome source. Run %s "
+                         " or call %s" % (
+                             self.fasta_type,
+                             self.genome_source.install_string_console(),
+                             self.genome_source.install_string_python()))
 
     def clear_cache(self):
         """Delete the local FASTA file and its associated database,
@@ -229,7 +213,8 @@ class SequenceData(object):
 
         if sequence_id not in self._sequence_cache:
             # all Ensembl identifiers start with ENS e.g. ENST, ENSP, ENSE
-            require_ensembl_id(sequence_id)
+            if require_ensembl_ids:
+                require_ensembl_id(sequence_id)
 
             # get sequence from database
             fasta_record = self.fasta_dictionary.get(sequence_id)
